Only rerender lens if state or text changed
If we rerender at every change, then when perform an undo, the
modification hook triggers, and we get a new lens, instead of just
using the lens that already was there. Now, when we undo, we don't
rerender and instead just use the lens that was already there, which
means that the undo history doesn't get cluttered.